FIX to allow ZPAQ to receive a block of data to compress that is greater than the block size used for compression.
This is a major overhaul of the way ZPAQ compresses.
Instead of using the libzpaq::compress(...) function, lrzip-next
will now use the StringBuffer class and the compressBlock function.

In addition, the computation of memory overhead for ZPAQ was
changed slightly per documentation. Adds up to 550MB per thread
to the overhead.

Would appreciate some more testing before moving this to whats-next
branch.